// packages/spark-bff/src/generated/typings/digitalBill.ts
export interface definitions {
// ...
  TransactionDetail: {
    /** The date of the transaction. */
    date: string;
    /** The type of transaction. */
    type:
      | 'Charged'
      | 'Late Fee'
      | 'Surcharge'
      | 'Credit'
      | 'Payment'
      | 'Failed';
    /** The value of the transaction. This be positive for debits and negative for credits. */
    value: number;
    /** The calculated balance (amount to pay) following this specific transaction. */
    calculatedBalance: number;
    /** Bill ID of the digital bill associated with this charge */
    billId?: string;
    /** A customer-friendly description of the transaction event. */
    description?: string;
  };
  /** Current charge with all possible breakdown in the bill. */
  CurrentChargeSummary: {
    /** The type of aggregated charge */
    type: string;
    /** The formatted NZD value of the charge */
    value: string;
  };
}